Check shouldAutoPublishInstall on auto publish.
Summary: We're not.  This was dropped during the port of the iOS implementation.  Thanks to Karsten Sperling (https://github.com/ksperling) for reporting.

Test Plan: modify a sample to change the value to false, and see publish not occur.  revert + delete the app & re-test.  see the publish occur.

Reviewers: mmarucheck

CC: msdkexp@, neko-dev@

Differential Revision: https://phabricator.fb.com/D574705

Blame Revision: D559309